+++ georiviere/river/tests/test_models.py
@@ -2,7 +2,8 @@ from django.conf import settings
 from django.contrib.gis.geos import Point, LineString
 from django.test import TestCase
 
-from georiviere.description.tests.factories import MorphologyFactory, StatusFactory
+from georiviere.finances_administration.tests.factories import AdministrativeFileFactory
+from georiviere.description.tests.factories import MorphologyFactory, StatusFactory, UsageFactory
 from georiviere.river.models import Stream
 from georiviere.river.tests.factories import TopologyFactory, StreamFactory
 
@@ -28,6 +29,10 @@ class StreamSourceLocationTest(TestCase):
             geom=LineString((10000, 10000), (50000, 50000)),
             source_location=Point(0, 0, srid=settings.SRID),
         )
+        cls.usage_point = UsageFactory.create(
+            geom=Point(10000, 10020)
+        )
+        cls.administrative_file = AdministrativeFileFactory.create()
 
     def test_source_location_default(self):
         """Test if source_location is set on save()"""
@@ -40,6 +45,11 @@ class StreamSourceLocationTest(TestCase):
         self.assertAlmostEqual(lat_min, -5.9838563092087576)
         self.assertAlmostEqual(lng_max, -1.0680441780204335)
         self.assertAlmostEqual(lat_max, -5.655019875165679)
+
+    def test_distance_to_source(self):
+        """Test distance from a given object to stream source according to differents geom"""
+        self.assertAlmostEqual(self.stream1.distance_to_source(self.usage_point), 20)
+        self.assertEqual(self.stream1.distance_to_source(self.administrative_file), None)
 
 
 class SnapTest(TestCase):